Create and use `Storage.kt` file instead of `storage` package.
