+++ client/src/main/kotlin/io/spine/examples/pingh/client/Storage.kt
@@ -24,8 +24,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.pingh.client.storage
+package io.spine.examples.pingh.client
 
+import com.google.common.annotations.VisibleForTesting
 import com.google.protobuf.Message
 import java.io.File
 import java.io.FileInputStream
@@ -34,6 +35,7 @@ import kotlin.io.path.Path
 import kotlin.io.path.createDirectories
 import kotlin.io.path.createFile
 import kotlin.io.path.exists
+import net.harawata.appdirs.AppDirsFactory
 
 /**
  * Stores data on disk in a sequence of bytes.
@@ -95,5 +97,68 @@ internal class FileStorage<T : Message>(location: FileLocation) {
      */
     internal fun clear() {
         FileOutputStream(file).close()
+    }
+}
+
+/**
+ * A location of a file on disk.
+ *
+ * @property dir The absolute path to the directory where the file is located.
+ * @property name The name of the file.
+ */
+internal class FileLocation private constructor(
+    internal val dir: String,
+    internal val name: String
+) {
+    internal companion object {
+        /**
+         * The path to application data within the user’s home directory.
+         */
+        private val appDirPath = AppDirPath.withoutVersion()
+
+        /**
+         * Creates a location of the file within the application's folder
+         * in the user data directory.
+         */
+        internal fun inAppDir(fileName: String): FileLocation =
+            FileLocation(appDirPath, fileName)
+    }
+}
+
+/**
+ * Deletes all files in the application's folder
+ * within the user data directory.
+ *
+ * For testing purposes only.
+ */
+@VisibleForTesting
+public fun clearAppDir() {
+    val path = AppDirPath.withoutVersion()
+    File(path).listFiles()?.forEach { file ->
+        if (file.isFile) {
+            file.delete()
+        } else {
+            file.deleteRecursively()
+        }
+    }
+}
+
+/**
+ * Provides the path to platform-specific application data
+ * within the user’s home directory.
+ */
+private object AppDirPath {
+    private const val author = "spine-examples"
+    private const val name = "Pingh"
+    private const val version = "1.0.0"
+
+    /**
+     * Returns the path to the application data,
+     * without including the current application version.
+     */
+    fun withoutVersion(): String {
+        val versionedPath = AppDirsFactory.getInstance()
+            .getUserDataDir(name, version, author)
+        return versionedPath.substring(0, versionedPath.length - version.length - 1)
     }
 }
